backend/requirements.txt
fastapi
uvicorn[standard]
aiogram
sqlmodel
asyncpg
python-dotenv
pydantic-settings
passlib[bcrypt]
bcrypt == 4.3.0
python-jose[cryptography]
aiosqlite